add start date add return full path report file

+++ red_api.py
@@ -11,16 +11,14 @@ mouth_report = today + relativedelta(months=-1)
 mouth_plan = today + relativedelta(months=+1)
 
 
-def put_api(red_api_key, gen_csv=False):
-    # - auth
+def put_api(red_api_key: object, gen_csv: object = False, start_date: object = None) -> object:
     red = redminelib.Redmine('http://red.eltex.loc', key=red_api_key)
-    if gen_csv:
+    if gen_csv and start_date is not None:
         # print(' auth +', red.url, '\n','Ваше Имя:', red.user.get('current')) #Получение адреса и имени пользователя
         # print(' Ваш ID:', red.user.get('current').firstname) # Получение ID пользователя
-
         # issue_custom_field_values_2 - Quality Assurance 6 статус это Rejected
         issues = red.issue.filter(cf_2='me', status_id='!6',
-                                  updated_on='>=' + str(mouth_report))
+                                  updated_on='>=' + str(start_date))
 
         # print_list_red(issues)
         shutil.rmtree('artifacts/', ignore_errors=True)
